Provide manpage for fssync-debug and most subcmds
This is another patch in the series to provide manpages for
dafs-related commands.

Most (if not all) documentation is coming from source
documentation and DAFS articles/presentations.

The primary motivation behind this is to comply with Debian
guidelines requiring all binaries to have respective manpages.

This patch has been sitting here for months, so it's good time
to finally submit it.

Note that fssync-debug also supports some vcg* commands, for
which I was not able to find any documentation I could use.
So, they are still left to be documented.

Change-Id: I9d067766e2a90b26646edc87f15a4f533c5c8d44
Reviewed-on: http://gerrit.openafs.org/2220
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>