windows  don't include assert.h (and afs support headers) in util_cr.c
Don't include afsconfig.h and afs/param.h in util_cr.c, nor
assert.h.  The assert was included but was in fact not used in
the file, and can be ommitted, per jaltman.  The OpenAFS
includes are then not needed.  An incorrect cast is also changed,
due to jaltman.

Change-Id: I3e1ad7a1d24473d6ad3ce2e401133b25b0a4c795
Reviewed-on: http://gerrit.openafs.org/2234
Reviewed-by: Jeffrey Altman <jaltman@openafs.org>
Tested-by: Jeffrey Altman <jaltman@openafs.org>