Removed kpasswd from openafs-file-list
The executable %{_bindir}/kpasswd was included in the
the 'openafs' base package.

This change removes the kpasswd executable from the base package.  It
is still built and is installed in the 'openafs-kpasswd' package.  I
believe this was the original intent of the 'openafs-kpasswd' package.

Reviewed-on: http://gerrit.openafs.org/2305
Reviewed-by: Simon Wilkinson <sxw@inf.ed.ac.uk>
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>
(cherry picked from commit b54c7ceefe884b57db76c929940afdfee86472e8)

Change-Id: Ib095073c090ff59c305defda3c7a3f1f34bcb8cd
Reviewed-on: http://gerrit.openafs.org/3753
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Jonathan Billings <jsbillin@umich.edu>
Reviewed-by: Derrick Brashear <shadow@dementia.org>