Linux: cache bypass: avoid unused variable warnings
Some variables are only used for AFS_64BIT_CLIENT code, so make
their declaration conditional as well.

Also, initialize 'code' while we're at it.  The compiler has a
legitimate complaint that it could be used before it's set,
for instance if we have AFS_64BIT_CLIENT but afs_serverHasNo64Bit
is true.

Change-Id: Ic04b07fffa9668123a9029bb42415a7c701c2339
Reviewed-on: http://gerrit.openafs.org/2317
Reviewed-by: Derrick Brashear <shadow@dementia.org>
Tested-by: Derrick Brashear <shadow@dementia.org>